move: SmsSignatureGenerator 파일을 util 폴더로 이동
