+++ apiserver/common/src/main/java/com/zipline/global/config/SmsWebClientConfig.java
@@ -1,7 +1,7 @@
 package com.zipline.global.config;
 
+import com.zipline.global.util.SmsSignatureGenerator;
 import java.util.Map;
-
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -9,32 +9,31 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import com.zipline.service.message.SmsSignatureGenerator;
 
 @Configuration
 public class SmsWebClientConfig {
 
-	@Value("${sms.api-key}")
-	String apiKey;
+  @Value("${sms.api-key}")
+  String apiKey;
 
-	@Bean
-	public WebClient webClient(SmsSignatureGenerator smsSignatureGenerator) {
-		Map<String, String> signatureResult = smsSignatureGenerator.generateSignature().block();
+  @Bean
+  public WebClient webClient(SmsSignatureGenerator smsSignatureGenerator) {
+    Map<String, String> signatureResult = smsSignatureGenerator.generateSignature().block();
 
-		if (signatureResult == null || signatureResult.isEmpty()) {
-			throw new IllegalArgumentException("유효하지 않은 signature 값 입니다.");
-		}
+    if (signatureResult == null || signatureResult.isEmpty()) {
+      throw new IllegalArgumentException("유효하지 않은 signature 값 입니다.");
+    }
 
-		String authorizationHeader = String.format(
-			"HMAC-SHA256 apiKey=%s, date=%s, salt=%s, signature=%s",
-			apiKey, signatureResult.get("time"), signatureResult.get("salt"),
-			signatureResult.get("hash")
-		);
+    String authorizationHeader = String.format(
+        "HMAC-SHA256 apiKey=%s, date=%s, salt=%s, signature=%s",
+        apiKey, signatureResult.get("time"), signatureResult.get("salt"),
+        signatureResult.get("hash")
+    );
 
-		return WebClient.builder()
-			.baseUrl("https://api.solapi.com/messages/v4")
-			.defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-			.defaultHeader(HttpHeaders.AUTHORIZATION, authorizationHeader)
-			.build();
-	}
+    return WebClient.builder()
+        .baseUrl("https://api.solapi.com/messages/v4")
+        .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+        .defaultHeader(HttpHeaders.AUTHORIZATION, authorizationHeader)
+        .build();
+  }
 }
